wrap client hydration in crypto support check
